requirements.txt
tensorflow
pillow
numpy
tf2onnx
onnxruntime
opencv-python
onnxconverter-common